Get survey, fill out survey, can't yet submit
