+++ client/smart-waterloo-web/src/pages/Survey/Survey.tsx
@@ -1,14 +1,17 @@
 import Navbar from "../../components/Navbar";
 import {useState, useContext, useEffect} from "react"
 import { MobileContext } from "../../App";
-import {exampleSurveys, SurveyDataType} from "../../data/Surveys";
+import { SurveyDataType} from "../../data/Surveys";
 import SurveyLanding from "./SurveyLanding";
 import "./Survey.css";
 import { useParams } from "react-router-dom";
 import SurveyQuestion from "./SurveyQuestion";
 import Cookies from "universal-cookie";
+import { getSurveyData } from "../../data/getData";
+import NotFound from "../NotFound";
 
 const defaultSurveyData:SurveyDataType = {
+	id: "",
 	name: "Loading...",
 	organization: "- - - - - - - - -",
 	length: "? mins",
@@ -22,7 +25,6 @@ const Survey = () => {
 	const { id } = useParams<"id">();
 	const {mobile} = useContext(MobileContext);
 	const [progress, setProgess] = useState(false);
-	const [surveyData, setSurveyData] = useState(defaultSurveyData);
 	const [loaded, setLoaded] = useState(false);
 	const [answers, setAnswers] = useState(defaultAnswers);
 	const greyText = {color: "grey"};
@@ -34,37 +36,42 @@ const Survey = () => {
 		newAnswers[index] = newVal;
 		setAnswers(newAnswers);
 	}
-	const getSurveyData = async () => {
-		setLoaded(true);
-		//TODO actual get 
-		console.log(id);
-		setTimeout(() => {
-			if (id && parseInt(id)<exampleSurveys.length)
-			setSurveyData(exampleSurveys[parseInt(id)]);
-			else alert("not found");
-		}, 1000)
+	const [notFound, setNotFound] = useState(false);
+	const [surveyData, setSurveyData] = useState({survey: defaultSurveyData, surveyDataSet: false});
+	const getSetSurveyData = async () => {
+		if (!id) return;
+		let {survey, success, errors} = await getSurveyData(id);
+		if (success) {
+			setSurveyData({ survey: survey, surveyDataSet: true })
+		} else {
+			setNotFound(true);
+			console.error(errors);
+		}
 	}
-	useEffect(() => {
-		if (!loaded) getSurveyData();
-	})
+	const [dataCalled, setDataCalled] = useState(false);
+	if (!dataCalled) {
+		getSetSurveyData();
+		setDataCalled(true);
+	}
 
-	if (answers.length !== surveyData.questions.length) {
+	if (answers.length !== surveyData.survey.questions.length) {
 		const newAnswers = [];
-		for (let i = 0; i<surveyData.questions.length; i++) newAnswers.push("");
+		for (let i = 0; i<surveyData.survey.questions.length; i++) newAnswers.push("");
 		setAnswers(newAnswers);
 	}
-	const questions = surveyData.questions.map((question, i) => {
+	const questions = surveyData.survey.questions.map((question, i) => {
 		return <SurveyQuestion key={i} index={i} answer={answers[i]} setParentAnswer={childSetAnswer} {...question}/>
 	})
 	const complete = answers.every(answer => answer!=="");
+	if (notFound || !id) return <NotFound />
 	return (
 		<>
 			<Navbar root={false}/>
 			<div className={"PageContainer"}>
 				<div className={mobile? "":"surveyPagePanel"}>
-					{mobile?<h4>{surveyData.name}</h4>:<h4 className={"surveyHeader"}>{surveyData.name}</h4>}
-					<p>{surveyData.organization}</p>
-					<p style={greyText}>{surveyData.length} to fill</p>
+					{mobile?<h4>{surveyData.survey.name}</h4>:<h4 className={"surveyHeader"}>{surveyData.survey.name}</h4>}
+					<p>{surveyData.survey.organization}</p>
+					<p style={greyText}>{surveyData.survey.length} to fill</p>
 					<div className={"surveyForm"}>
 						
 					</div>
