+++ client/smart-waterloo-web/src/pages/Surveys/Surveys.tsx
@@ -3,7 +3,7 @@ import Navbar from "../../components/Navbar";
 import { MobileContext, IdContext, AddressContext } from "../../App";
 import {useContext, useState} from "react";
 import "./Surveys.css";
-import {defaultSurveysState, exampleSurveys} from "../../data/Surveys";
+import {defaultSurveysState} from "../../data/Surveys";
 import SurveyPanel from "./SurveyPanel";
 import { useNavigate } from "react-router-dom";
 import { defaultAccountData} from "../../data/account";
